Notas

+++ app/src/main/java/ipvc/estg/cidadeinteligente/ActivityNotes.kt
@@ -1,9 +1,14 @@
 package ipvc.estg.cidadeinteligente
 
 import android.app.Activity
+import android.app.AlertDialog
+import android.content.DialogInterface
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import android.text.TextWatcher
+import android.view.LayoutInflater
+import android.widget.EditText
 import android.widget.TextView
 import android.widget.Toast
 import androidx.lifecycle.Observer
@@ -14,6 +19,8 @@ import androidx.recyclerview.widget.RecyclerView
 import com.google.android.material.floatingactionbutton.FloatingActionButton
 import ipvc.estg.cidadeinteligente.adapter.NoteAdapter
 import ipvc.estg.cidadeinteligente.viewmodel.NoteViewModel
+import kotlinx.android.synthetic.main.alert_edit.view.*
+import kotlinx.android.synthetic.main.recyclerline.*
 
 
 class ActivityNotes : AppCompatActivity(),NoteAdapter.OnDeleteClickListener, NoteAdapter.OnUpdateClickListener{
@@ -55,20 +62,70 @@ class ActivityNotes : AppCompatActivity(),NoteAdapter.OnDeleteClickListener, Not
                 noteViewModel.insert(note)
 
 
-        } else{
-            Toast.makeText(
-                applicationContext,
-                "Nota nao inserida",
-                Toast.LENGTH_LONG).show()
         }
     }
 
 
     override fun onDeleteClick(id: Int) {
-        noteViewModel.deleteNote(id)
+        val yes: String = getString(R.string.yes)
+        val no: String = getString(R.string.no)
+        val title: String = getString(R.string.title)
+        val dia:String = getString(R.string.confirm)
+
+        val builder = AlertDialog.Builder(this)
+                .setTitle(title)
+                .setMessage(dia)
+
+        builder.setPositiveButton(yes,DialogInterface.OnClickListener { dialog, with ->
+            noteViewModel.deleteNote(id)
+            dialog.cancel()
+        })
+
+        builder.setNegativeButton(no, DialogInterface.OnClickListener { dialog, with ->
+            dialog.cancel()
+        })
+
+        val alert = builder.create()
+        alert.show()
+
     }
 
     override fun onUpdateClick(id: Int, titulo: String, notes: String) {
-        Toast.makeText(this, "$id ,$titulo ,$notes", Toast.LENGTH_SHORT).show()
+        val title: String = getString(R.string.edit)
+
+        val mDialogView = LayoutInflater.from(this).inflate(R.layout.alert_edit, null)
+
+        val mBuiler = AlertDialog.Builder(this)
+                .setView(mDialogView)
+                .setTitle(title)
+
+        mDialogView.titulo_edit.setText(titulo)
+        mDialogView.nota_edit.setText(notes)
+
+        val mAlertDialog = mBuiler.show()
+
+        mDialogView.save.setOnClickListener {
+            val tituloUpdated = mDialogView.titulo_edit.text.toString()
+            val notesUpdated = mDialogView.nota_edit.text.toString()
+
+            val error:String = getString(R.string.error)
+
+            if(tituloUpdated.length == 0) {
+                mDialogView.titulo_edit.setError(error)
+            }
+            if(notesUpdated.length == 0){
+                mDialogView.nota_edit.setError(error)
+            }
+            if(tituloUpdated.length != 0 && notesUpdated.length != 0){
+                mAlertDialog.dismiss()
+                noteViewModel.updateNote(id, tituloUpdated, notesUpdated)
+            }
+        }
+
+        mDialogView.cancel.setOnClickListener {
+            mAlertDialog.dismiss()
+        }
+
     }
+
 }
